+++ .prettierrc.js
@@ -1,0 +1,4 @@
+export default {
+  // Solana Kit has very long names and prettier defaults to 80 characters
+  printWidth: 120,
+};