+++ src/lib/messages.ts
@@ -6,8 +6,8 @@ export const signMessageFromWalletApp = async (
 ): Promise<string> => {
   const base58Decoder = getBase58Decoder();
   const encodedMessage = new TextEncoder().encode(message);
-  // Oddly, there's only a modifyAndSignMessages (which accepts an array of messages and returns an array of results) 
-  // but no singular modifyAndSignMessage. 
+  // Oddly, there's only a modifyAndSignMessages (which accepts an array of messages and returns an array of results)
+  // but no singular modifyAndSignMessage.
   // TODO: should be fixed upstream.
   const results = await messageSigner.modifyAndSignMessages([
     {
@@ -15,12 +15,12 @@ export const signMessageFromWalletApp = async (
       signatures: {},
     },
   ]);
-  const result = results[0]
+  const result = results[0];
 
   // Get the first (and should be only) signature from the result
   const signature = Object.values(result?.signatures)[0];
   if (!signature) {
-    throw new Error('Could not find signature in the result');
+    throw new Error("Could not find signature in the result");
   }
   return base58Decoder.decode(signature);
-}
+};