+++ src/lib/smart-transactions.ts
@@ -1,13 +1,12 @@
 // Based on smart-transaction.ts from https://github.com/mcintyre94/helius-smart-transactions-web3js2
 
 import {
-  IInstruction,
-  getComputeUnitEstimateForTransactionMessageFactory,
+  Instruction,
   appendTransactionMessageInstruction,
   TransactionMessage,
   isWritableRole,
   isInstructionWithData,
-  CompilableTransactionMessage,
+  TransactionMessageWithFeePayer,
   createSolanaRpcFromTransport,
   sendAndConfirmTransactionFactory,
   TransactionWithBlockhashLifetime,
@@ -15,15 +14,19 @@ import {
   Commitment,
   SOLANA_ERROR__TRANSACTION_ERROR__ALREADY_PROCESSED,
   isSolanaError,
+  SOLANA_ERROR__JSON_RPC__SERVER_ERROR_SEND_TRANSACTION_PREFLIGHT_FAILURE,
+  assertIsTransactionWithinSizeLimit,
+  Transaction,
 } from "@solana/kit";
 import {
   getSetComputeUnitPriceInstruction,
   identifyComputeBudgetInstruction,
   COMPUTE_BUDGET_PROGRAM_ADDRESS,
   ComputeBudgetInstruction,
+  estimateComputeUnitLimitFactory,
 } from "@solana-program/compute-budget";
 import { getAbortablePromise } from "@solana/promises";
-import { DEFAULT_TRANSACTION_RETRIES, DEFAULT_TRANSACTION_TIMEOUT } from "./constants";
+import { DEFAULT_TRANSACTION_RETRIES, DEFAULT_TRANSACTION_TIMEOUT, SECONDS } from "./constants";
 
 export const getPriorityFeeEstimate = async (
   rpc: ReturnType<typeof createSolanaRpcFromTransport>,
@@ -33,7 +36,7 @@ export const getPriorityFeeEstimate = async (
 ): Promise<number> => {
   const accountKeys = [
     ...new Set([
-      ...transactionMessage.instructions.flatMap((instruction: IInstruction) =>
+      ...transactionMessage.instructions.flatMap((instruction: Instruction) =>
         (instruction.accounts ?? [])
           .filter((account) => isWritableRole(account.role))
           .map((account) => account.address),
@@ -77,7 +80,7 @@ export const getPriorityFeeEstimate = async (
 
 export const getComputeUnitEstimate = async (
   rpc: ReturnType<typeof createSolanaRpcFromTransport>,
-  transactionMessage: CompilableTransactionMessage,
+  transactionMessage: TransactionMessage & TransactionMessageWithFeePayer,
   abortSignal: AbortSignal | null = null,
 ) => {
   // add placeholder instruction for CU price if not already present
@@ -94,27 +97,35 @@ export const getComputeUnitEstimate = async (
     ? transactionMessage
     : appendTransactionMessageInstruction(getSetComputeUnitPriceInstruction({ microLamports: 0n }), transactionMessage);
 
-  const computeUnitEstimateFn = getComputeUnitEstimateForTransactionMessageFactory({ rpc });
-  // TODO: computeUnitEstimateFn expects an explicit 'undefined' for abortSignal,
+  const estimateComputeUnitLimit = estimateComputeUnitLimitFactory({ rpc });
+  // TODO: estimateComputeUnitLimit expects an explicit 'undefined' for abortSignal,
   // fix upstream
-  return computeUnitEstimateFn(transactionMessageToSimulate, {
+  return estimateComputeUnitLimit(transactionMessageToSimulate, {
     abortSignal: abortSignal ?? undefined,
   });
 };
 
 export const sendTransactionWithRetries = async (
   sendAndConfirmTransaction: ReturnType<typeof sendAndConfirmTransactionFactory>,
-  transaction: FullySignedTransaction & TransactionWithBlockhashLifetime,
+  transaction: Transaction & FullySignedTransaction & TransactionWithBlockhashLifetime,
   options: {
     maximumClientSideRetries: number;
     abortSignal: AbortSignal | null;
     commitment: Commitment;
+    timeout?: number | null;
   } = {
     maximumClientSideRetries: DEFAULT_TRANSACTION_RETRIES,
     abortSignal: null,
     commitment: "confirmed",
+    timeout: null,
   },
 ) => {
+  if (options.commitment === "finalized") {
+    console.warn(
+      "Using finalized commitment for transaction with retries is not recommended. This will likely result in blockhash expiration.",
+    );
+  }
+
   let retriesLeft = options.maximumClientSideRetries;
 
   const transactionOptions = {
@@ -127,10 +138,31 @@ export const sendTransactionWithRetries = async (
     maxRetries: 0n,
   };
 
+  let timeout: number;
+  if (options.timeout) {
+    timeout = options.timeout;
+  } else {
+    switch (options.commitment) {
+      case "processed":
+        timeout = 5 * SECONDS;
+        break;
+      case "confirmed":
+        timeout = 15 * SECONDS;
+        break;
+      case "finalized":
+        timeout = 30 * SECONDS;
+        break;
+      default:
+        timeout = DEFAULT_TRANSACTION_TIMEOUT;
+        break;
+    }
+  }
+
   while (retriesLeft) {
     try {
+      assertIsTransactionWithinSizeLimit(transaction);
       const txPromise = sendAndConfirmTransaction(transaction, transactionOptions);
-      await getAbortablePromise(txPromise, AbortSignal.timeout(DEFAULT_TRANSACTION_TIMEOUT));
+      await getAbortablePromise(txPromise, AbortSignal.timeout(timeout));
       break;
     } catch (error) {
       if (error instanceof DOMException && error.name === "TimeoutError") {
@@ -141,6 +173,12 @@ export const sendTransactionWithRetries = async (
         // race condition where the transaction is processed between throwing the
         // `TimeoutError` and our next retry
         break;
+      } else if (isSolanaError(error, SOLANA_ERROR__JSON_RPC__SERVER_ERROR_SEND_TRANSACTION_PREFLIGHT_FAILURE)) {
+        if (error.cause && isSolanaError(error.cause, SOLANA_ERROR__TRANSACTION_ERROR__ALREADY_PROCESSED)) {
+          // race condition where the transaction is processed between throwing the
+          // `TimeoutError` and our next retry and our simulation fails
+          break;
+        }
       } else {
         throw error;
       }