+++ src/lib/transactions.ts
@@ -1,13 +1,14 @@
 import {
   appendTransactionMessageInstructions,
   assertIsTransactionMessageWithSingleSendingSigner,
+  assertIsTransactionWithinSizeLimit,
   Commitment,
   createSolanaRpcFromTransport,
   createTransactionMessage,
   getBase58Decoder,
   getBase58Encoder,
   getSignatureFromTransaction,
-  IInstruction,
+  Instruction,
   KeyPairSigner,
   pipe,
   sendAndConfirmTransactionFactory,
@@ -16,6 +17,8 @@ import {
   signAndSendTransactionMessageWithSigners,
   signTransactionMessageWithSigners,
   TransactionSendingSigner,
+  assertIsTransactionMessageWithBlockhashLifetime,
+  Blockhash,
 } from "@solana/kit";
 import { getComputeUnitEstimate, getPriorityFeeEstimate, sendTransactionWithRetries } from "./smart-transactions";
 import { getSetComputeUnitLimitInstruction, getSetComputeUnitPriceInstruction } from "@solana-program/compute-budget";
@@ -50,7 +53,7 @@ export const sendTransactionFromInstructionsWithWalletAppFactory = (
     abortSignal = null,
   }: {
     feePayer: TransactionSendingSigner;
-    instructions: Array<IInstruction>;
+    instructions: Array<Instruction>;
     abortSignal?: AbortSignal | null;
   }) => {
     const { value: latestBlockhash } = await rpc.getLatestBlockhash().send({ abortSignal });
@@ -58,11 +61,7 @@ export const sendTransactionFromInstructionsWithWalletAppFactory = (
       createTransactionMessage({ version: 0 }),
       (message) => setTransactionMessageFeePayerSigner(feePayer, message),
       (message) => setTransactionMessageLifetimeUsingBlockhash(latestBlockhash, message),
-      (message) =>
-        appendTransactionMessageInstructions(
-          instructions,
-          message,
-        ),
+      (message) => appendTransactionMessageInstructions(instructions, message),
     );
     assertIsTransactionMessageWithSingleSendingSigner(transactionMessage);
     const signatureBytes = await signAndSendTransactionMessageWithSigners(transactionMessage);
@@ -71,7 +70,6 @@ export const sendTransactionFromInstructionsWithWalletAppFactory = (
   };
   return sendTransactionFromInstructionsWithWalletApp;
 };
-
 
 export const sendTransactionFromInstructionsFactory = (
   rpc: ReturnType<typeof createSolanaRpcFromTransport>,
@@ -87,19 +85,25 @@ export const sendTransactionFromInstructionsFactory = (
     skipPreflight = true,
     maximumClientSideRetries = enableClientSideRetries ? DEFAULT_TRANSACTION_RETRIES : 0,
     abortSignal = null,
+    timeout,
   }: {
     feePayer: KeyPairSigner;
-    instructions: Array<IInstruction>;
+    instructions: Array<Instruction>;
     commitment?: Commitment;
     skipPreflight?: boolean;
     maximumClientSideRetries?: number;
     abortSignal?: AbortSignal | null;
+    timeout?: number;
   }) => {
-    const { value: latestBlockhash } = await rpc.getLatestBlockhash().send({ abortSignal });
+    // use a placeholder for simulation so we can wait to do the getLatestBlockhash call as the last step
+    let placeholderBlockhash = {
+      blockhash: "11111111111111111111111111111111" as Blockhash,
+      lastValidBlockHeight: 0n,
+    } as const;
 
     let transactionMessage = pipe(
       createTransactionMessage({ version: 0 }),
-      (message) => setTransactionMessageLifetimeUsingBlockhash(latestBlockhash, message),
+      (message) => setTransactionMessageLifetimeUsingBlockhash(placeholderBlockhash, message),
       (message) => setTransactionMessageFeePayerSigner(feePayer, message),
       (message) => appendTransactionMessageInstructions(instructions, message),
     );
@@ -124,7 +128,12 @@ export const sendTransactionFromInstructionsFactory = (
       );
     }
 
+    const { value: latestBlockhash } = await rpc.getLatestBlockhash().send({ abortSignal });
+    transactionMessage = setTransactionMessageLifetimeUsingBlockhash(latestBlockhash, transactionMessage);
+    assertIsTransactionMessageWithBlockhashLifetime(transactionMessage);
+
     const signedTransaction = await signTransactionMessageWithSigners(transactionMessage);
+    assertIsTransactionWithinSizeLimit(signedTransaction);
 
     const signature = getSignatureFromTransaction(signedTransaction);
 
@@ -134,6 +143,7 @@ export const sendTransactionFromInstructionsFactory = (
           maximumClientSideRetries,
           abortSignal,
           commitment,
+          timeout,
         });
       } else {
         await sendAndConfirmTransaction(signedTransaction, {
@@ -143,10 +153,12 @@ export const sendTransactionFromInstructionsFactory = (
       }
     } catch (thrownObject) {
       const error = thrownObject as ErrorWithTransaction;
-
+      // getTransaction does not support processed commitment so use confirmed instead
+      // https://solana.com/docs/rpc/http/gettransaction
+      const commitmentToUse = commitment === "processed" ? "confirmed" : commitment;
       const transaction = await rpc
         .getTransaction(signature, {
-          commitment,
+          commitment: commitmentToUse,
           maxSupportedTransactionVersion: 0,
         })
         .send();